Organise files for clarity
