+++ packages/types/src/types/requests/index.ts
@@ -4,5 +4,5 @@
  *
  */
 
-export { MeditrakSurveyResponseRequest } from './MeditrakSurveyResponseRequest';
-export { DataTablePreviewRequest } from './DataTablePreviewRequest';
+export { MeditrakSurveyResponseRequest } from './central-server/MeditrakSurveyResponseRequest';
+export { DataTablePreviewRequest } from './data-table-server/DataTablePreviewRequest';
